agile-backend/requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==2.3.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.5
greenlet==3.3.0
itsdangerous==2.2.0
Jinja2==3.1.2
joblib==1.5.0
MarkupSafe==2.1.3
numpy==1.24.4
pandas==1.5.3
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
SQLAlchemy==2.0.45
threadpoolctl==3.6.0
typing_extensions==4.15.0
Werkzeug==3.1.5
